+++ internal/core/note_find.go
@@ -11,10 +11,10 @@ import (
 
 // NoteFindOpts holds a set of filtering options used to find notes.
 type NoteFindOpts struct {
-	// Filter used to match the notes with FTS predicates.
+	// Filter used to match the notes with the given MatchStrategy.
 	Match opt.String
-	// Search for exact occurrences of the Match string.
-	ExactMatch bool
+	// Text matching strategy used with Match.
+	MatchStrategy MatchStrategy
 	// Filter by note hrefs.
 	IncludeHrefs []string
 	// Filter excluding notes at the given hrefs.
@@ -160,4 +160,30 @@ func NoteSorterFromString(str string) (NoteSorter, error) {
 	}
 
 	return sorter, nil
+}
+
+// MatchStrategy represents a text matching strategy used when filtering notes with `--match`.
+type MatchStrategy int
+
+const (
+	// Full text search.
+	MatchStrategyFts MatchStrategy = iota + 1
+	// Exact text matching.
+	MatchStrategyExact
+	// Regular expression.
+	MatchStrategyRe
+)
+
+// MatchStrategyFromString returns a MatchStrategy from its string representation.
+func MatchStrategyFromString(str string) (MatchStrategy, error) {
+	switch str {
+	case "fts", "f", "":
+		return MatchStrategyFts, nil
+	case "re", "grep", "r":
+		return MatchStrategyRe, nil
+	case "exact", "e":
+		return MatchStrategyExact, nil
+	default:
+		return 0, fmt.Errorf("%s: unknown match strategy\ntry fts (full-text search), re (regular expression) or exact", str)
+	}
 }